refactor(card): weight, height 타입 변경

+++ src/main/java/com/dekk/card/application/command/CardCreateCommand.java
@@ -11,7 +11,7 @@ public record CardCreateCommand(
     String originId,
     Boolean isActive,
     Platform platform,
-    Integer height,
-    Integer weight
+    Double height,
+    Double weight
 ) {
 }
