+++ src/main/java/com/dekk/card/domain/model/Card.java
@@ -56,9 +56,9 @@ public class Card {
     @Column(nullable = false)
     private Platform platform;
 
-    private Integer height;
+    private Double height;
 
-    private Integer weight;
+    private Double weight;
 
     private Card(
             CardImage cardImage,
@@ -66,8 +66,8 @@ public class Card {
             String originId,
             Boolean isActive,
             Platform platform,
-            Integer height,
-            Integer weight
+            Double height,
+            Double weight
     ) {
         this.cardImage = cardImage;
         this.tagIds = tagIds;
